Forenklet task til å alltid ferdigstille åpne GodkjenneVedtak-oppgaver og opprette BehandleSak-oppgave. Kjører kun task dersom åpen oppgave ikke er BehandleSak eller BehandleUnderkjentVedtak.
